Stop SHA errors related to: Token Length Exception
Some websites and providers generate a secret key that technically is not the proper length to use in generating a OTP. If you pad from the left when such a secret key is detected you can fix the key length without altering the output. This is needed for example with websites like Amazon that currently return a 52 character secret key.